Refactor bundle manager janitor.
